requirements.txt
pandas
numpy
scikit-learn
matplotlib
fastapi
python-box #config box
pyYAML
ensure
joblib
types-PyYAML
seaborn
streamlit
ipykernel > 6.0.0
pip-chill==1.0.1
requests
category_encoders
apache-airflow
pyarrow
fastparquet
boto3<=1.15.3
python-dotenv
docker
azureml-core
great_expectations

-e .